+++ d3-examples/dag/index.html
@@ -15,7 +15,7 @@
         window.grafo = grafo
       })();
 
-      const store = new Map();
+      // const store = new Map();
 
       function draw(raw) {
         const dag = d3.dratify()(raw);
