+++ package.json
@@ -38,6 +38,8 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "browser-sync": "^2.26.3",
+    "css-loader": "^1.0.1",
+    "style-loader": "^0.23.1",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10"
   }
