+++ src/index.html
@@ -4,15 +4,15 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>🙏git-bless</title>
+  <title>🙏 git-bless</title>
 </head>
 <body>
-  <svg width="300px" height="150px">
+  <!-- <svg width="300px" height="150px">
     <rect x="20" y="20" width="20px" height="20" rx="5" ry="5" />
     <rect x="60" y="20" width="20px" height="20" rx="5" ry="5" />
     <rect x="100" y="20" width="20px" height="20" rx="5" ry="5" />
-  </svg>
-
+  </svg> -->
+  
   <div id="app"></div>
 </body>
 </html>
