+++ webpack.config.js
@@ -18,14 +18,20 @@ module.exports = {
     host: '0.0.0.0'
   },
   module: {
-    rules: [{
-      test: /\.js$/,
-      exclude: [/node_modules/],
-      use: [{
-        loader: 'babel-loader',
-        options: { presets: ['es2015', 'react'] },
-      }],
-    }]
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: [/node_modules/],
+        use: [{
+          loader: 'babel-loader',
+          options: { presets: ['es2015', 'react'] },
+        }],
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader']
+      }
+    ]
   },
   plugins: [HtmlWebpackPluginConfig]
 }
